Rev 5
Prompt: there is no need to store duplicate previously played videos. if the same video is in the list, just remove the older one. i would also like the link text to be the name of the video, not the url. also, please add a thumbnail for each video in the list.